[PR-17] User 생성, 위치 업데이트, 경로 추천 받아오기 API 연동

+++ android/app/src/main/java/kr/pnu/ga2019/domain/repository/UserInfoRepository.kt
@@ -12,8 +12,8 @@ interface UserInfoRepository : Repository {
 
     fun updateCurrentLocation(
         memberPk: Int,
-        locationX: Double,
-        locationY: Double
+        locationX: Int,
+        locationY: Int
     ): Completable
 
     fun getAllUserLocation(memberPk: Int): Single<List<User>>
